Clone serialization toolkit for each user
